split cases.tests
